implemented view cabang

+++ SIRETAIL/src/main/java/apap/tugasakhir/SIRETAIL/service/CabangService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface CabangService {
     void addCabang(CabangModel cabang);
     List<CabangModel> getListCabang();
-    CabangModel getCabangByIdCabang(Long idCabang);
+    CabangModel getCabangByIdCabang(Integer idCabang);
 }
